restructuring and modularization
