+++ haskell/src/Flint.hs
@@ -6,14 +6,11 @@ import Web.Scotty
 import Flint.Types (Config (..))
 import Flint.Index (index) 
 import Flint.Blog
+import Flint.Utils
 import Network.Wai.Middleware.Static
 import Network.Wai.Middleware.RequestLogger
 import Network.Wai.Middleware.Gzip
 import Control.Monad.IO.Class
-import Web.Scotty.Internal.Types qualified
-
-type Scotty = ScottyM ()
-type Action = ActionM ()
 
 lucid :: Html () -> Action
 lucid = html . renderText
@@ -23,15 +20,12 @@ routes config = do
   get "/" do
     lucid $ index config Nothing
 
-  get "/hc" do
-    pure ()
-
   get "/blog" do
     lucid $ index config Nothing
     
   get "/blog/:article" do
     articleId <- param "article"
-    articles <- liftIO $ getArticles config
+    articles <- liftIO $ getMarkdownWithMeta config parseArticle "/blog/"
     let meta = head [article.meta | article <- articles, article.slug == articleId]
     lucid $ index config $ Just meta
